Improve forms no-JS support

+++ app/routes/labels/$labelId/delete.tsx
@@ -1,24 +1,20 @@
-import {
-  Box,
-  Button,
-  Checkbox,
-  FormControlLabel,
-  Typography,
-} from '@mui/material';
+import { Box, Button, Title } from '@mantine/core';
 import { Label } from '@prisma/client';
 import { withZod } from '@remix-validated-form/with-zod';
 import { useState } from 'react';
 import {
   ActionFunction,
-  Form,
   LoaderFunction,
   MetaFunction,
   json,
   redirect,
   useLoaderData,
 } from 'remix';
+import { useHydrated } from 'remix-utils';
+import { ValidatedForm } from 'remix-validated-form';
 import { z } from 'zod';
 import { zfd } from 'zod-form-data';
+import ValidatedCheckbox from '~/components/ValidatedCheckbox';
 import { extractIntFromParam } from '~/lib/helpers.server';
 import { ensureAuthenticated } from '~/lib/middleware.server';
 import { prisma } from '~/lib/prisma.server';
@@ -35,26 +31,30 @@ const paramsSchema = zfd.formData({
   labelId: zfd.numeric(z.number().min(0)),
 });
 
-const formSchema = withZod(z.object({}));
+const formSchema = withZod(
+  z.object({
+    confirm: zfd.checkbox(),
+  }),
+);
 
 const responseSchema = z.null();
 
 export const outputSchema = formActionResponseSchema(responseSchema);
 
-/*
- * Delete a label.
- *
- * URL parameters:
- *   labelId: number  The id of the label to delete
- */
+// Delete a label
 export const action: ActionFunction = async (actionArgs) =>
   validatedFormAction({
     actionArgs,
     paramsSchema,
     formSchema,
     responseSchema,
-    async action({ request, data: { labelId } }) {
+    async action({ request, data: { confirm, labelId } }) {
       const userId = await ensureAuthenticated(request);
+
+      if (!confirm) {
+        // The user didn't check the confirm checkbox, so do nothing
+        return null;
+      }
 
       // Delete the label
       const { count } = await prisma.label.deleteMany({
@@ -101,11 +101,14 @@ export const loader: LoaderFunction = async ({ request, params }) => {
 export default function DeleteLabelRoute() {
   const { label } = useLoaderData<LabelData>();
 
+  const isHydrated = useHydrated();
+
   const [confirmDelete, setConfirmDelete] = useState<boolean>(false);
 
   return (
     <Box
-      component={Form}
+      component={ValidatedForm}
+      validator={formSchema}
       method="post"
       sx={{
         width: '25em',
@@ -114,22 +117,21 @@ export default function DeleteLabelRoute() {
         gap: '1em',
       }}
     >
-      <Typography variant="h3" component="h2">
-        Delete label
-      </Typography>
-      <FormControlLabel
-        control={
-          <Checkbox
-            checked={confirmDelete}
-            onChange={(event) => setConfirmDelete(event.target.checked)}
-          />
-        }
+      <Title order={2}>Delete label</Title>
+      <ValidatedCheckbox
+        name="confirm"
         label={
           `Are you sure you want to delete the label "${label.name}"? ` +
           `This action cannot be undone.`
         }
+        checked={confirmDelete}
+        onChange={(event) => setConfirmDelete(event.target.checked)}
       />
-      <Button type="submit" color="error" disabled={!confirmDelete}>
+      <Button
+        type="submit"
+        color="error"
+        disabled={isHydrated && !confirmDelete}
+      >
         Delete
       </Button>
     </Box>
